refactor: Update budget settings related forms/hooks to use centralized hook
