+++ js/src/domains/settings/grants/components/GrantForm.tsx
@@ -3,7 +3,9 @@ import { Stack, TextField } from '@mui/material';
 import { useSafeIntl } from 'bluesquare-components';
 import InputComponent from 'Iaso/components/forms/InputComponent';
 import { useTranslatedErrors } from 'Iaso/libs/validation';
+import { useGetBudgetSettings } from '../../../../hooks/useGetBudgetSettings';
 import { useGetExtendedFormikContext } from '../../../../hooks/useGetExtendedFormikContext';
+import { getCurrencySymbol } from '../../../../utils/currency';
 import { MESSAGES } from '../../../messages';
 import { GrantFormValues } from '../types';
 import { DonorSelect } from './DonorSelect';
@@ -13,6 +15,8 @@ export const GrantForm: FC = () => {
 
     const { values, errors, touched, setFieldValueAndState } =
         useGetExtendedFormikContext<GrantFormValues>();
+    const { data: budgetSettings } = useGetBudgetSettings();
+    const currencySymbol = getCurrencySymbol(budgetSettings?.local_currency);
 
     const getErrors = useTranslatedErrors({
         errors,
@@ -54,6 +58,10 @@ export const GrantForm: FC = () => {
                     labelString={formatMessage(MESSAGES.grantAmount)}
                     withMarginTop={false}
                     wrapperSx={{ flexGrow: 1 }}
+                    numberInputOptions={{
+                        decimalScale: 2,
+                        prefix: currencySymbol,
+                    }}
                 />
             </Stack>
             <DonorSelect
