+++ js/src/domains/settings/interventions/components/InterventionFormWrapper.tsx
@@ -11,6 +11,7 @@ import { setNestedObjectValues } from 'formik';
 import { DeleteRestoreModal } from 'Iaso/components/DeleteRestoreModals/DeleteRestoreModal';
 import { CardStyled } from '../../../../components/CardStyled';
 import { SettingsFormContainer } from '../../../../components/styledComponents';
+import { useGetBudgetSettings } from '../../../../hooks/useGetBudgetSettings';
 import { ExtendedFormikProvider } from '../../../../hooks/useGetExtendedFormikContext';
 import { useGetMetricTypes } from '../../../dataLayers/hooks/useGetMetrics';
 import { useDeleteIntervention } from '../../../interventions/hooks/useDeleteIntervention';
@@ -26,7 +27,6 @@ import {
 import { MESSAGES } from '../../../messages';
 import { useGetGrants } from '../../grants/hooks/useGetGrants';
 import { InterventionProvider } from '../contexts/InterventionContext';
-import { useGetBudgetSettings } from '../hooks/useGetBudgetSettings';
 import { useInterventionFormState } from '../hooks/useInterventionFormState';
 import { InterventionFormValues } from '../types/interventionForm';
 import { InterventionBasicForm } from './InterventionBasicForm';
